Move activity and event resource actions to timetracking domain
