+++ lib/omedis_web/live/activity_live/index.ex
@@ -204,7 +204,7 @@ defmodule OmedisWeb.ActivityLive.Index do
   defp apply_action(socket, :edit, %{"id" => id}) do
     actor = socket.assigns.current_user
     organisation = socket.assigns.organisation
-    activity = Activity.by_id!(id, actor: actor, tenant: organisation)
+    activity = TimeTracking.get_activity_by_id!(id, actor: actor, tenant: organisation)
 
     if Ash.can?({activity, :update}, actor, tenant: organisation) do
       socket
@@ -261,7 +261,7 @@ defmodule OmedisWeb.ActivityLive.Index do
 
   defp list_paginated_activities(socket, params) do
     PaginationUtils.list_paginated(socket, params, :activities, fn offset ->
-      Activity.list_paginated(
+      TimeTracking.list_paginated_activities(
         %{group_id: socket.assigns.group.id},
         actor: socket.assigns.current_user,
         page: [count: true, offset: offset],
