+++ lib/omedis_web/live/activity_live/show.ex
@@ -4,7 +4,7 @@ defmodule OmedisWeb.ActivityLive.Show do
   alias Omedis.Accounts.Organisation
   alias Omedis.Groups.Group
   alias Omedis.Projects.Project
-  alias Omedis.TimeTracking.Activity
+  alias Omedis.TimeTracking
 
   on_mount {OmedisWeb.LiveHelpers, :assign_and_broadcast_current_organisation}
 
@@ -126,7 +126,12 @@ defmodule OmedisWeb.ActivityLive.Show do
     organisation = socket.assigns.organisation
     group = Group.by_slug!(group_slug, actor: socket.assigns.current_user, tenant: organisation)
     groups = Ash.read!(Group, actor: socket.assigns.current_user, tenant: organisation)
-    activity = Activity.by_id!(id, actor: socket.assigns.current_user, tenant: organisation)
+
+    activity =
+      TimeTracking.get_activity_by_id!(id,
+        actor: socket.assigns.current_user,
+        tenant: organisation
+      )
 
     next_position = activity.position
